requirements.txt
pydantic>=2.0
crewai
python-dotenv
faiss-cpu
numpy